pyproject.toml
[tool.poetry]
name = "e7-gear-calc"
version = "0.1.0"
description = ""
authors = ["Dylan Jung <dj5822@gmail.com>"]
readme = "README.md"
packages = [{include = "e7_gear_calc"}]

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.2.0"
pytesseract = "^0.3.10"
opencv-python = "^4.6.0.66"
numpy = "^1.23.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
project = "e7_gear_calc:main"
